# Add a stored procedure for PostgreSQL rlm_sqlippool.