--- a/raddb/mods-config/sql/ippool/postgresql/queries.conf
+++ b/raddb/mods-config/sql/ippool/postgresql/queries.conf
@@ -37,6 +37,17 @@ allocate_find = "\
 #	FOR UPDATE SKIP LOCKED"
 
 #
+#  Use a stored procedure to find the address. This requires PostgreSQL >= 9.5 as
+#  SKIP LOCKED is used.
+#
+# allocate_find = "\
+#  SELECT find_previous_or_new_framedipaddress( \
+#  	'%{control:${pool_name}}', \
+#  	'%{SQL-User-Name}', \
+#  	'%{Calling-Station-Id}' \
+#  	)"
+
+#
 #  If you prefer to allocate a random IP address every time, use this query instead
 #  Note: This is very slow if you have a lot of free IPs.
 #