This matches the functionality of the existing example query which returns the users previous IP, if possible, however it runs approx 400x faster.

I have a modest test postgresql server, with 4GB RAM, and a radippool table with 2x65k address pools.

I have been testing with the queries to assign addresses. One example query assigns the user the previous address they had, and a new address if not. This is *very* slow, because of how the ORDER BY clause is formed. I have written a postgresql function and it vastly improves performance.

I have been testing a query which assigns simply the least recently used address, which I refer to here as "give me any address":
```
allocate_find = "\
  SELECT framedipaddress FROM ${ippool_table} \
  WHERE pool_name = '%{control:${pool_name}}' AND expiry_time < 'now'::timestamp(0) \
  ORDER BY expiry_time \
  LIMIT 1 \
  FOR UPDATE SKIP LOCKED"
```
I have a variant of this without ordering, as well.

A single execution of the "give me any address" query was taking approx 0.2-0.4ms on my system, and the "give me the previous address" was taking approx 145ms.

Running in a loop 1000x shows the same order of magnitude difference - however not quite the same. I believe this is because the DB caches the queries - I'm not completely sure.

Original query:
```
radius=# DO $$BEGIN
radius$#   FOR i IN 1..1000 LOOP
radius$#     PERFORM framedipaddress FROM radippool WHERE pool_name = 'CustomersNat-ipv4' AND expiry_time < 'now'::timestamp(0) ORDER BY (username <> 'test'), (callingstationid <> 'test'), expiry_time LIMIT 1 FOR UPDATE SKIP LOCKED;
radius$#   END LOOP;
radius$# END; $$;
DO
Time: 64665.157 ms (01:04.665)
```

New query:
```
radius=# DO $$BEGIN
radius$#   FOR i IN 1..1000 LOOP
radius$#     PERFORM find_previous_or_new_framedipaddress('CustomersNat-ipv4', 'test', 'test');
radius$#   END LOOP;
radius$# END; $$;
DO
Time: 162.592 ms
```

"Give me any address" query:
```
radius=#
radius=# DO $$BEGIN
radius$#   FOR i IN 1..1000 LOOP
radius$#     PERFORM framedipaddress FROM radippool WHERE pool_name = 'CustomersNat-ipv4' AND expiry_time < 'now'::timestamp(0) LIMIT 1 FOR UPDATE SKIP LOCKED;
radius$#   END LOOP;
radius$# END; $$;
DO
Time: 36.211 ms
```

"Give me any address" query with ordering:
```
radius=# DO $$BEGIN
radius$#   FOR i IN 1..1000 LOOP
radius$#     PERFORM framedipaddress FROM radippool WHERE pool_name = 'CustomersNat-ipv4' AND expiry_time < 'now'::timestamp(0) ORDER BY expiry_time LIMIT 1 FOR UPDATE SKIP LOCKED;
radius$#   END LOOP;
radius$# END; $$;
DO
Time: 70.142 ms
```

I have done testing with FreeRADIUS with these queries. I have a NAS simulator python script which simulates 100,000 users with 50 user/s disconnect/reconnects. My test RADIUS server runs in a different datacenter (approx 2ms away) from the database server.

I got the following times for getting 100,000 users online:

Old query: couldn't. 50 user/s churn was too much
New query: approx 2m30s
"Give me any address" query with ordering: approx 1m40s
"Give me any address" query without ordering: approx 6m

Oddly, asking the DB to order and give the "least recently used" address is faster over time than just giving any address. I'm not entirely clear why this is. Both start out handling approx 1000/s, and over time the unordered query degrades to around 500/s or so, from memory.

Either way, the new query I am submitting here is significantly faster than the previous one, to the point of being usable.

If it's interesting I can also include my ordered and unordered "Any Address" queries.