@nward 2019-02-28
I should note some key details here that I omitted;
 - in smaller deployments, the performance benefit here is greatly reduced. The performance issue comes largely because sorting by the results of boolean expressions is not indexable. If there are fewer expired (i.e. "free") addresses to sort then there is a much smaller penalty when sorting them.
 - these build on top of my previous "SKIP LOCKED" changes. I have not tested these without SKIP LOCKED. They'll work - but will be slower than without it.
- my tests are against a PostgreSQL database fronted by PgPool, with asynchronous streaming replication.
- I am running with "max_servers" 30, sql pool with max of 32, and my simulation runs locally and generates packets "from" 10 different NASes.