@alandekok 2019-02-28
If you could sign the commits that would be good.  Otherwise I'll pull them over manually.

Also, My $0.02 would be to put the schema into `raddb/mods-config/sql/ippool/postgresql/procedure.sql` .  That way it doesn't have to be commented out.